requirements.txt
#
# list of dependencies for Polymer
#
# you can install them with pip using:
# $ pip install -r requirements.txt

cython
numpy
scipy
python-hdf4
netcdf4
glymur    # for S2
lxml      # for S2
pyproj    # for S2
pandas    # for Level1_ASCII
pyepr     # for MERIS
ecmwf-api-client   # for Ancillary_ERA (download ERA-Interim files)
pygrib             # for Ancillary_ERA
